Update Spanish.po PB 0.74.3
Translation up to Public Build 0.754.3

Changes in the previous version:

I have removed the bars for the genre because they mess up the text a lot and it becomes more difficult to read. I have tried to make it as neutral as possible.

The proper names are untranslated, except Dragon Matriarch and Queen Bee, as they are actually titles. It makes it easy to recognize the characters when they talk about them.
Mastah and Master (as Fern and Blossom call the breeder) are also untranslated to avoid gender.

The names of the races and variants are all in lower case for the dialogues. There are a lot of them, and although it is capitalized to highlight important information, the complete data appears on the task board.

Reinvented Sugar walls before balls, for a text that rhymes although the meaning remains unclear. If I find something better I will change it in the next update.

I think that with these changes it is cleaner. For anything contact me on the Discord server.